+++ packages/gatsby/gatsby-plugin-clipanion-cli/gatsby-node.js
@@ -80,8 +80,8 @@ exports.sourceNodes = ({actions, createNodeId, createContentDigest}, opts) => {
       const contentDigest = createContentDigest(content);
 
       createNode({
-        ...commands[t],
-        id: createNodeId(t),
+        ...command,
+        id: createNodeId(command.path),
         parent: null,
         children: [],
         internal: {