+++ packages/yarnpkg-builder/sources/tools/makeConfig.ts
@@ -31,7 +31,7 @@ export function findTsconfig() {
   return npath.join(currTsContextRoot, `tsconfig.json`);
 }
 
-// @ts-ignore: @types/webpack-merge depends on @types/webpack, which isn't compatible with the webpack 5 types
+// @ts-expect-error: @types/webpack-merge depends on @types/webpack, which isn't compatible with the webpack 5 types
 export const makeConfig = (config: webpack.Configuration): webpack.Configuration => merge(identity<webpack.Configuration>({
   mode: `none`,
   devtool: false,