+++ packages/yarnpkg-doctor/sources/cli.ts
@@ -25,6 +25,7 @@ async function findFiles(pattern: string, cwd: PortablePath, {ignoredFolders = [
     absolute: true,
     cwd: npath.fromPortablePath(cwd),
     ignore: [`**/node_modules/**`, ...ignoredFolders.map(p => `${npath.fromPortablePath(p)}/**`)],
+    gitignore: true,
   });
 
   return files.map(p => {
@@ -196,7 +197,7 @@ async function buildJsonNode(p: PortablePath, accesses: Array<string>) {
     /*setParentNodes */ true,
   );
 
-  // @ts-ignore
+  // @ts-expect-error
   let node: ts.Node = sourceFile.statements[0].expression;
   if (!node)
     throw new Error(`Invalid source tree`);